IMPALA-5509: Runtime filter : Extend runtime filter to support Dictionary values
If a runtime filter uses only a sigle column, then it can be used to
filter Parquet dictionaries, and if all dictionary values are filtered,
out, the whole row group can be skipped.

The chance of false positives grow exponentially with the size of the
dictionary, so this optimisation is only useful for small dictionaries.

So this optimization has several criteria:
- the dictionary size of the join column must be small (but not 1,
  as min/max based filtering can solve that case), otherwise
  there is a large chance of false positives
- there must be no matching value in the row group, which is
  unlikely if the whole table has several matching rows, and
  those are shuffled randomly

I did not find any good candidate in tpc-h, so I had create an artifical
test case to demonstrate the optimization. The concept is to choose a small
dimension table (nation) and a larger fact table (supplier), and remove
every row  from the fact table that references a certain row in the
dimension table. If a filter in the join only allows this row, we can be
sure that the build side will only contain keys that are not present in
the dictionary. As the size of the dimension table is small, we can
be also sure that the dictionary will be small, so false positives
are unlikely to cause a problem.

Another way would be to use a fact table where the join column is sorted,
so the dictionaries if  individual row groups would be smaller.

Testing:

use tpch_parquet;

create table supplier2 stored as parquet
as select supplier.* from supplier, nation where s_nationkey != 4;

select * from supplier2 join nation
on s_nationkey = n_nationkey where n_name = "EGYPT";

Check the query profile of the select - without the patch,
NumDictFilteredRowGroups will be 0. With the patch, it will
be 1 (in the scan nodes of table supplier2).

TODO:
- skip this filtering if the dictionary is above a certain size
- currently only one runtime filter per column is supported ( I am not
sure whether multiple runtime filters per column makes sense or not)
- maybe add a counter the measure the effect of this filter exclusively

Change-Id: I5cf42800efa1204214a4f308c4b61273613b3222